Added management of pkill of processes A,C,P,Q

+++ src/Analyzer/A.c
@@ -1,38 +1,59 @@
 #include "../lib/lib.h"
 
-process *p;  //Declaring p (it's global because hendle_sigint can't have parameters, only int sig)
+int value_return = 0;  //Valore di ritorno, globale per "send_to_R"
+process *p;            //Declaring p (it's global because hendle_sigint can't have parameters, only int sig)
+int fd_fifo;           //pipe fifo con R
 
 void handle_sigint(int sig) {
-    printf("\n[!] Ricevuta terminazione, inizio terminazione processi ... \n");
+    printf("\n[!] Ricevuta terminazione da A, inizio terminazione processi C,P,Q ... \n");
     int i = p->count - 1;  //start from the end
-    while (i != 0)         //while we haven't controlled every single process
-    {
-        if (p->pid[i] > 0)  //Processo padre
+    if (i > 0) {
+        while (i != 0)  //while we haven't controlled every single process
         {
-            if (kill(p->pid[i], 9) == 0) {                                     //Tries to kill process with pid saved in pid[i]
-                printf("\tProcesso %d terminato con successo!\n", p->pid[i]);  //if it success you terminated it correctly
-            } else {
-                printf("\t[!] Errore, non sono riuscito a chiudere il processo %d!", p->pid[i]);  //if it fail something is wrong
+            if (p->pid[i] > 0)  //Processo padre
+            {
+                if (kill(p->pid[i], 9) == 0) {                                     //Tries to kill process with pid saved in pid[i]
+                    printf("\tProcesso %d terminato con successo!\n", p->pid[i]);  //if it success you terminated it correctly
+                } else {
+                    printf("\t[!] Errore, non sono riuscito a chiudere il processo %d!", p->pid[i]);  //if it fail something is wrong
+                }
             }
+            i--;  //i-- otherwise it will go to infinity
         }
-        i--;  //i-- otherwise it will go to infinity
+    }
+    if (!close(fd_fifo))  //close fifo
+    {
+        printf("[!] Chiusura fifo completata\n");
     }
     freeList(p);  //free memory allocated for p
-    printf("[!] ... Chiusura processi terminata\n");
+    printf("[!] ... Chiusura processi C,P,Q terminata\n");
     exit(-1);  //return exit with error -1
+}
+
+void sig_term_handler(int signum, siginfo_t *info, void *ptr) {
+    value_return = err_kill_process_A();
+}
+
+void catch_sigterm() {
+    static struct sigaction _sigact;
+
+    memset(&_sigact, 0, sizeof(_sigact));
+    _sigact.sa_sigaction = sig_term_handler;
+    _sigact.sa_flags = SA_SIGINFO;
+
+    sigaction(SIGTERM, &_sigact, NULL);
 }
 
 int main(int argc, char *argv[]) {
     printf("Start A\n");
+    catch_sigterm();
     signal(SIGINT, handle_sigint);  //Handler for SIGINT (Ctrl-C)
 
     p = create_process(1);  //Allocate dynamically p with dimension 1
 
-    int value_return = 0;  //Valore di ritorno, globale per "send_to_R"
-
     //COMMUNICATION WITH R
     const char *fifo = "/tmp/A_R_Comm";  //Nome fifo con R
-    int fd_fifo;                         //pipe fifo con R
+    //int fd_fifo;                         //pipe fifo con R
     char print_method[DIM_CMD];
     int retrieve = TRUE;
     int _write_val = -1;
@@ -148,7 +169,7 @@ int main(int argc, char *argv[]) {
                 //_close = TRUE;
                 if (!_close) {
                     if (read(STDIN_FILENO, cmd, DIM_CMD) > 0) {
-                        if(!strncmp(cmd, "close", 5)) {
+                        if (!strncmp(cmd, "close", 5)) {
                             _close = TRUE;
                         }
                     }
@@ -180,7 +201,7 @@ int main(int argc, char *argv[]) {
                             //Settare l'output
                         }
                         //Aggiungere flags
-                        
+
                         retrieve = TRUE;
                         do {
                             _write_val = write(fd_fifo, tmp_resp, DIM_RESP);
